# Documentation/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gradox2 API",
        "description": "API de gestión de archivos académicos para la aplicación Gradox2",
        "contact": {
            "name": "Gradox2 Team",
            "email": "gradox2app@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/uploadProposal/upload": {
            "post": {
                "tags": [
                    "Propuestas de Archivos"
                ],
                "summary": "Proponer archivo",
                "description": "Sube un archivo para su revisión mediante votación. Se especifica el nivel de visibilidad (PUBLIC, RESTRICTED, PRIVATE)",
                "operationId": "uploadProposal",
                "parameters": [
                    {
                        "name": "fileProposalRequest",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/UploadFileProposalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propuesta creada, pendiente de votación",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileProposalResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o límite de subidas pendientes alcanzado"
                    }
                }
            }
        },
...
        "/uploadProposal": {
            "get": {
                "tags": [
                    "Propuestas de Archivos"
                ],
                "summary": "Listar propuestas",
                "description": "Lista todas las propuestas de archivos, con soporte de paginación y ordenación",
                "operationId": "getAllUploadProposals",
                "parameters": [
                    {
                        "name": "paged",
                        "in": "query",
                        "description": "Activar paginación",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Tamaño de página (máx 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Campo de ordenación (id, quorumRequired, approvalThreshold, status, actionType, createdAt, closedAt)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista devuelta",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autenticado"
                    }
                }
            }
        },
        "/uploadProposal/{id}": {
            "get": {
                "tags": [
                    "Propuestas de Archivos"
                ],
                "summary": "Detalle de propuesta",
                "description": "Obtiene los datos de una propuesta de archivo por ID",
                "operationId": "getUploadProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propuesta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos devueltos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileProposalResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propuesta no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Propuestas de Archivos"
                ],
                "summary": "Borrar propuesta",
                "description": "Elimina una propuesta pendiente (solo el proponente puede borrarla)",
                "operationId": "deleteUploadProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propuesta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propuesta eliminada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "La propuesta no está pendiente o no eres el proponente"
                    },
                    "404": {
                        "description": "Propuesta no encontrada"
                    }
                }
            }
        },
        "/uploadProposal/{id}/download": {
            "get": {
                "tags": [
                    "Propuestas de Archivos"
                ],
                "summary": "Descargar archivo propuesto",
                "description": "Genera una URL firmada para descargar el archivo temporal asociado a una propuesta pendiente (válida 120 segundos)",
                "operationId": "downloadProposalFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propuesta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL firmada de descarga generada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Propuesta o archivo no encontrado"
                    }
                }
            }
        },
...
        "/files/{id}": {
            "get": {
                "tags": [
                    "Archivos"
                ],
                "summary": "Detalle de archivo",
                "description": "Obtiene los metadatos de un archivo publicados (visibilidad aplicada sobre el uploader)",
                "operationId": "getFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos devueltos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Archivos"
                ],
                "summary": "Proponer borrado",
                "description": "Crea una propuesta de eliminación para un archivo publicado (requiere votación)",
                "operationId": "requestFileDeletion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propuesta de borrado creada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileProposalResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ya existe una propuesta de borrado activa"
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                }
            }
        },
        "/files/{id}/download": {
            "get": {
                "tags": [
                    "Archivos"
                ],
                "summary": "Descargar archivo",
                "description": "Genera una URL firmada para descargar el archivo directamente desde el bucket (válida 120 segundos)",
                "operationId": "downloadFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL firmada de descarga generada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FileResponse": {
                "type": "object",
                "description": "Respuesta con los datos de un archivo publicado",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID del archivo",
                        "example": 1
                    },
                    "fileName": {
                        "type": "string",
                        "description": "Nombre del archivo",
                        "example": "Apuntes Tema 1"
                    },
                    "description": {
                        "type": "string",
                        "description": "Descripción del archivo",
                        "example": "Apuntes completos del tema 1"
                    },
                    "fileType": {
                        "type": "string",
                        "description": "Tipo de archivo (APUNTES, EJERCICIO, EXAMEN)",
                        "enum": [
                            "APUNTES",
                            "EXAMEN",
                            "EJERCICIO",
                            "OTRO"
                        ]
                    },
                    "subject": {
                        "type": "string",
                        "description": "Nombre de la asignatura",
                        "example": "Matemáticas"
                    },
                    "uploaderUsername": {
                        "type": "string",
                        "description": "Nombre del uploader (según visibilidad puede ser 'anonymous')",
                        "example": "aliceuser"
                    },
                    "visibilityLevel": {
                        "type": "string",
                        "description": "Nivel de visibilidad (PUBLIC, RESTRICTED, PRIVATE)",
                        "enum": [
                            "PUBLIC",
                            "RESTRICTED",
                            "PRIVATE"
                        ],
                        "example": "PUBLIC"
                    },
                    "score": {
                        "type": "number",
                        "format": "double",
                        "description": "Puntuación media (0.0 - 10.0)",
                        "example": 8.5
                    }
                }
            },
...
            "UploadFileProposalRequest": {
                "type": "object",
                "description": "Solicitud de propuesta de subida de archivo (multipart/form-data)",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título del archivo",
                        "example": "Apuntes Tema 1",
                        "maxLength": 150,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "description": "Descripción del archivo",
                        "example": "Apuntes completos del tema 1",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "type": {
                        "type": "string",
                        "description": "Tipo de archivo (APUNTES, EJERCICIO, EXAMEN)",
                        "enum": [
                            "APUNTES",
                            "EXAMEN",
                            "EJERCICIO",
                            "OTRO"
                        ]
                    },
                    "subjectId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la asignatura",
                        "example": 1
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Archivo a subir (PDF, máx 10 MB)"
                    },
                    "visibilityLevel": {
                        "type": "string",
                        "description": "Nivel de visibilidad del uploader (PUBLIC, RESTRICTED, PRIVATE)",
                        "enum": [
                            "PUBLIC",
                            "RESTRICTED",
                            "PRIVATE"
                        ],
                        "example": "PUBLIC"
                    },
                    "validFile": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "file",
                    "subjectId",
                    "type"
                ]
            },
            "FileProposalResponse": {
                "type": "object",
                "description": "Respuesta con los datos de una propuesta de archivo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la propuesta",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "Título del archivo propuesto",
                        "example": "Apuntes Tema 1"
                    },
                    "description": {
                        "type": "string",
                        "description": "Descripción del archivo",
                        "example": "Apuntes completos del tema 1"
                    },
                    "proposer": {
                        "type": "string",
                        "description": "Nombre del proponente (según visibilidad puede ser 'anonymous')",
                        "example": "aliceuser"
                    },
                    "visibilityLevel": {
                        "type": "string",
                        "description": "Nivel de visibilidad solicitado (PUBLIC, RESTRICTED, PRIVATE)",
                        "example": "PUBLIC"
                    },
                    "subjectName": {
                        "type": "string",
                        "description": "Nombre de la asignatura",
                        "example": "Matemáticas"
                    },
                    "course": {
                        "type": "string",
                        "description": "Nombre del curso",
                        "example": "1º ESO"
                    },
                    "type": {
                        "type": "string",
                        "description": "Tipo de archivo (APUNTES, EJERCICIO, EXAMEN)",
                        "example": "APUNTES"
                    },
                    "status": {
                        "type": "string",
                        "description": "Estado (PENDING, APPROVED, REJECTED)",
                        "example": "PENDING"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}